% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files_to_server.R
\name{create_dir_str}
\alias{create_dir_str}
\title{Creates the directory structure on the shiny server}
\usage{
create_dir_str(shiny_server_pth, url, dirs)
}
\arguments{
\item{shiny_server_pth}{string. A string indicating the path to the shiny server.}

\item{url}{string. A string indicating which directory to place the files to.}

\item{dirs}{vector. A vector of directories to create on the shiny server.}
}
\description{
Creates the directory structure on the shiny server
}
