% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files_to_server.R
\name{deploy_app}
\alias{deploy_app}
\title{Deploy app}
\usage{
deploy_app(url, shiny_server_pth = "/srv/shiny-server")
}
\arguments{
\item{url}{string. A string indicating which directory to place the files to.}

\item{shiny_server_pth}{string. A string indicating the path to the shiny server.}
}
\description{
Copies folder and file structure to the shiny server
}
