% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files_to_server.R
\name{deploy_main}
\alias{deploy_main}
\title{Copy files}
\usage{
deploy_main(deploy)
}
\arguments{
\item{deploy}{list. A list containing the urls and run values for development and production.}
}
\description{
Copy files
}
