% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files_to_server.R
\name{remove_old_fls}
\alias{remove_old_fls}
\title{Remove old files}
\usage{
remove_old_fls(shiny_server_pth, url, fls, ignore_fl_server)
}
\arguments{
\item{shiny_server_pth}{string. A string indicating the path to the shiny 
server.}

\item{url}{string. A string indicating which directory to place the files to.}

\item{fls}{vector. A list of files to copy to the shiny server.}
}
\description{
Remove old files
}
