% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files_to_server.R
\name{test_config}
\alias{test_config}
\title{Test config}
\usage{
test_config(deploy)
}
\arguments{
\item{deploy}{list. A list containing the urls and run values for development and production.}
}
\description{
Test config
}
