% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_files_to_server.R
\name{create_fl_str}
\alias{create_fl_str}
\title{Create fl structure on the shiny server}
\usage{
create_fl_str(shiny_server_pth, url, fls)
}
\arguments{
\item{shiny_server_pth}{string. A string indicating the path to the shiny server.}

\item{url}{string. A string indicating which directory to place the files to.}

\item{fls}{vector. A list of files to copy to the shiny server.}
}
\description{
Create fl structure on the shiny server
}
